+++ Example/ImageStreamerApp.swiftpm/Package.swift
@@ -52,3 +52,4 @@ let package = Package(
     ],
     swiftLanguageModes: [.version("6")]
 )
+